local server stress test setup
