+++ packages/test/local-server-stress-tests/src/test/localServerStressOrderSequentially.spec.ts
@@ -104,6 +104,7 @@ describe("Local Server Stress with rollback", () => {
 		},
 		skip: [
 			...[97], // channel maps should be the same size
+			...[30, 32, 35, 67], // task manager - The number of tasks queues are not the same
 		],
 	});
 });
